Check if docker is not installed before installing. Remove edge channel.

+++ stage2/02-cloud-init/01-run.sh
@@ -6,6 +6,12 @@ install -m 644 files/meta-data ${ROOTFS_DIR}/boot/
 on_chroot << EOF
 echo 'Symlinking cloud-init nocloud-net seed files to /boot partition'
 mkdir -p /var/lib/cloud/seed/nocloud-net
-ln -s /boot/user-data /var/lib/cloud/seed/nocloud-net/user-data
-ln -s /boot/meta-data /var/lib/cloud/seed/nocloud-net/meta-data
+if [ ! -f /boot/user-data ]
+then
+  ln -s /boot/user-data /var/lib/cloud/seed/nocloud-net/user-data
+fi
+if [ ! -f /boot/meta-data ]
+then
+  ln -s /boot/meta-data /var/lib/cloud/seed/nocloud-net/meta-data
+fi
 EOF
